+++ os-sim/www/riskmeter/resetip.php
@@ -1,10 +1,18 @@
 <?php
+    require_once 'classes/Security.inc';
 
+    $ip = GET('ip');
+
+    ossim_valid($ip, OSS_IP_ADDR, 'illegal:'._("IP address"));
+
+    if (ossim_error()) {
+        die(ossim_error());
+    }
+                                
     require_once ('ossim_db.inc');
     require_once ('classes/Host_qualification.inc');
     $db = new ossim_db();
     $conn = $db->connect();
-    $ip = mysql_escape_string($_GET["ip"]);
 
     Host_qualification::delete($conn, $ip);
 ?>
